refactor: Extract format_size to shared utils module
Move duplicate format_size function from status.py and gc.py to a new
shared utility module at magpie.utils.formatting. This addresses PR #140
review feedback about code duplication.
