+++ src/magpie/utils/__init__.py
@@ -1,0 +1,7 @@
+"""Shared utility functions for magpie."""
+
+from __future__ import annotations
+
+from magpie.utils.formatting import format_size
+
+__all__ = ["format_size"]
