+++ src/magpie/utils/formatting.py
@@ -1,0 +1,22 @@
+"""Formatting utilities for human-readable output."""
+
+from __future__ import annotations
+
+
+def format_size(size_bytes: int) -> str:
+    """Format byte size as human-readable string.
+
+    Args:
+        size_bytes: Size in bytes.
+
+    Returns:
+        Human-readable size string (e.g., "1.5 MB").
+    """
+    if size_bytes < 1024:
+        return f"{size_bytes} B"
+    elif size_bytes < 1024 * 1024:
+        return f"{size_bytes / 1024:.1f} KB"
+    elif size_bytes < 1024 * 1024 * 1024:
+        return f"{size_bytes / (1024 * 1024):.1f} MB"
+    else:
+        return f"{size_bytes / (1024 * 1024 * 1024):.1f} GB"
